create hex <-> rgb helpers
