+++ src/util/stringManipulation.tsx
@@ -57,3 +57,19 @@ export function pluralize(count: number | boolean, singular: string, plural?: st
     return singular + 's';
 }
 
+export const hexToRgb = (hex: string) => {
+    const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+    return result ? {
+        r: parseInt(result[1], 16),
+        g: parseInt(result[2], 16),
+        b: parseInt(result[3], 16)
+    } : { r: 54, g: 162, b: 235 }; // fallback blue
+};
+
+export const rgbToHex = (r: number, g: number, b: number) => {
+    const toHex = (value: number) => {
+        const hex = value.toString(16);
+        return hex.length === 1 ? '0' + hex : hex;
+    };
+    return `#${toHex(r)}${toHex(g)}${toHex(b)}`;
+};
